requirements.txt
asgiref==3.2.7
Django==3.0.6
django-tinymce4-lite==1.8.0
gunicorn==20.0.4
jsmin==2.2.2
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
